pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zomato-recommendation"
version = "0.1.0"
description = "AI-powered restaurant recommendations (Zomato-style dataset + LLM)"
requires-python = ">=3.11"
dependencies = [
    "datasets>=2.14.0",
    "fastapi>=0.115.0",
    "groq>=0.11.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.39.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["zomato_recommendation*"]

[tool.setuptools.package-data]
zomato_recommendation = ["config/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
